combined generate prime code into generate_keys.py
