+++ RSA/__init__.py
@@ -1,2 +1,2 @@
-from generate_prime import generate_prime
+from generate_keys import generate_keys
 from encrypt_decrypt import encrypt, decrypt
